Added in Navbar, and created basics for the rest of the homepage
