+++ src/pages/Contact.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+
+const Contact = () => {
+  return (
+    <div className="font-sans text-white bg-gray-900 min-h-screen flex flex-col items-center justify-center text-center px-4">
+      <div className="max-w-screen-xl mx-auto p-8 rounded-lg bg-glass">
+        <h1 className="text-4xl md:text-6xl leading-tight font-bold mb-4">Contact Me</h1>
+        <p className="text-lg">This is the Contact Me section.</p>
+      </div>
+    </div>
+  );
+};
+
+export default Contact;
